Explicit store for tests
Summary:
Previously we used a global test. While this was fine, if we
accidently ran the test in our staging or prod servers, it would write
objects to the transaction log, that could be very error prone
(especially if it deletes some objects but not all). We now make it so
that if there's already a store open, then we won't be able to run the
test at all.

TODO: we need to check that all objects are deleted: t234

Fixes t233

Maniphest Tasks: T233

Differential Revision: https://phabricator.tdrhq.com/D4984

GitOrigin-RevId: f9e2837f61fd599d30cd4726512a1352c6bcc35f