Low hanging fruit: cache timestamps
Summary:
Surprisingly, rendering timestamps are actually pretty
expensive. More expensive than looking it up from a precomputed
cache. That's what we're doing here.

The cache doesn't get cleaned at the moment. My assumption is that
there's a fixed number of timestamps that we render, so there's a
limit to how much this can grow.

After this, all of the optimizations that are left are going to be in
markup.

Differential Revision: https://phabricator.tdrhq.com/D5196

GitOrigin-RevId: feaa1d50f7e550164942dc340426f995b9753614