Simplify the test handlers for put-file
Summary:
I'm not able to pinpoint the bug in the tests, but it appears
that with-raw-post-data-as-tmp file doesn't do the right thing here,
and the tests are failing at least on my laptop. They seem fine on
desktop.

I don't know what's up. I'll link back to this diff from the
with-raw-post-data-as-tmp-file docs

Differential Revision: https://phabricator.tdrhq.com/D5631

GitOrigin-RevId: 565a228a3e030c0be5413768b455ff2f1b054d78