Store the previous screenshot as part of the iterator closure
Summary:
Previously, we were recomputing the screenshot. This was
fine, but if we're going to do something more complex by using the
screenshot name of the previous screenshot instead of just the
hardcoded screenshot-name, it makes more sense to save the previously
generated screenshot as part of the closure. (Otherwise, we'll have to
do ugly buggy things like modifying `screenshot-name`).

Differential Revision: https://phabricator.tdrhq.com/D5790

GitOrigin-RevId: 9f19be8d52a0b16578323fab6b91b7e7c4398f67