Switch verification to use app, except...
Summary:
The automated flow from the webhook doesn't work correctly.

The reason for this is that the hook for repositories-added and
installation-id management happen at the same time, and we don't know
which one gets run first.

In the next step, we'll make sure that the installation-id hook calls
a dedicated hook for when new repos are added.

Differential Revision: https://phabricator.tdrhq.com/D6318

GitOrigin-RevId: 9b54a4e34452e46a14d2869cbc8abf9b74af68d7