Inversion of control for templates
Summary:
Is this the best solution? Another solution is, instead of returning a markup tag, return an object of type say, TEMPLATED-MARKUP.

Then it would be the responsibility of the request dispatch to render that markup with template.

This other solution is purely functional. Might be more error prone with things like modifying the returned HTML.

This solution is a more incremental solution.

Differential Revision: https://phabricator.tdrhq.com/D6327

GitOrigin-RevId: 777f1f5a2c1420d15c2bc49adf5708c20bccf7a2