Avoid urlencoding the entire form data
Summary:
This speeds up file uploads. Not as much as I would want to,
but from 19s per 20MB to 14s per 20MB. CPU goes from 12s to 5s.

Arguably, the gains from optimizing the CPU at this point are
marginal, so I'm going to stop optimizing. It was just a curious
distraction to see how far I can optimize this.

Differential Revision: https://phabricator.tdrhq.com/D6415

GitOrigin-RevId: 8bbe21b8aaeb59325336131350b06ba3b33b3cb7