Simplify the response from http-get
Summary:
Previously we made it return a hash-table to keep it
compatible with dexador, but we've switched our loyalties back to
drakma, so this simplifies the interface.

PLUS: An alist is a lot more natural here than using a hash-table, and
it's less prone since I can immutably add and remove headers.

Differential Revision: https://phabricator.tdrhq.com/D6460

GitOrigin-RevId: 162c739ec0ba0fabdfe3b25733aa4ff686c443ad