Refactor out logic for abstract-pr-promoter
Summary:
Previously, we were using this logic from GitHub for both
BitBucket and GitLab. By moving this logic to an abstract promoter, we
can test the edge cases independently of the GitHub/BitBucket/GL API
specifics.

Test Plan:
The tests should cover a good amount of logic for BitBucket
and GitHub. GitLab is a little iffy. I'll test what I can in staging,
and then test in prod and revert if something is off.

Differential Revision: https://phabricator.tdrhq.com/D6565

GitOrigin-RevId: 01f2166c093d30881d8cf00d610047a77d270cbd