Add logging to verify that transaction log was loaded without errors
Summary:
This is the most common case, so we want to make sure we have
a nice message right in our faces when everything went well.

Fixes T583

Test Plan:
Ran `verify --store ...` on staging, and made sure the good
message was shown. For the bad case, I'm relying on the unit tests to
happy path the warning.

Maniphest Tasks: T583

Differential Revision: https://phabricator.tdrhq.com/D7053

GitOrigin-RevId: 607e88ed4d71f1b918e396ea0bdbb1c3e811b706