Don't use deprecated hash functions from OpenSSL
Summary:
In particular this breaks on LibreSSL on Mac.

Previously, this error wasn't showing up because we were loading
cl+ssl, which in turn loaded OpenSSL from homebrew, which had the
SHA/MD5 hash functions.

But when we got rid of cl+ssl from the SDK, we were unable to build
the SDK for mac. Hopefully this fixes it.

Build @all

Differential Revision: https://phabricator.tdrhq.com/D7249

GitOrigin-RevId: 05c7b6ba5ef7f7be2fb6f3aad89c4c5dbbe57fc1