Prevent ~users requests
Summary:
This isn't really a security flaw at the moment, because we
don't let the the actual data be requested.

But because of Lispwork's pathname handling this *reveals* whether a
certain user exists or not on the server. If it crashes, the user does
not exist, if it 404s then the user exists. This could potentially
trigger some automated tests by attackers.

Fixes T843

Maniphest Tasks: T843

Differential Revision: https://phabricator.tdrhq.com/D8049

GitOrigin-RevId: 1aa55b09dbc70529172cf2b6e7db8aa1d747cd54