Handle api-key being unbound
Summary: See T929. The root cause seems to be that our initial api-key deletion code just made the api-key slot unbound. I'm not sure why I did that, the comment is not very helpful but hints that deleting it could be tricky, so I'll put that off for another time.

Differential Revision: https://phabricator.tdrhq.com/D8260

GitOrigin-RevId: 149ccefe2fc718c37d8f3bf85e8f5b8f811262dc