Make nibbles an acceptor mixin, instead of a plugin
Summary:
Nibbles have grown to be a fundamental architecture
component. Leaving it as a plugin makes it awkward to access, since
other modules will need access to the *nibble-plugin* object.

It's a lot more natural to keep the nibbles a fundamental part of the
acceptor itself, so we make it a mixin.

(Warning: can this be hot reloaded? not sure.)

Differential Revision: https://phabricator.tdrhq.com/D8488

GitOrigin-RevId: 7629763539febe03e2199b8b915d091607f7148c