Add support for notifying master and merge base branches
Summary:
Fixes T1088.

No GK for this change, the only way to revert is to revert the commit.

Test Plan:
This is tricky to test. The main behavioral changes from the code, are
* When the main branch is same as the work branch. Previously, we were most likely not sending a change, but it's potentially possible that we were sending a red
* Unchanged runs, which currently only affect T*****.
* merge queue, which currently only affect T*****.

My plan:
* On staging, test with https://bitbucket.org/tdrhq/fast-example

Pushing to master resulted in:

{F177693}

Pushing to PR resulted in:

{F177695}

I was able to accept the changes.

* Push to prod, test with https://github.com/tdrhq/monorepo-example.

In both cases, we have to create a PR, see the changes look reasonable. Merge the PR, see that main branch has been updated.

Maniphest Tasks: T1088

Differential Revision: https://phabricator.tdrhq.com/D9167

GitOrigin-RevId: 8ad7330ef0b9eb90169a31b0a81f0086281df5aa