Initial attempt at forwarding requests
Summary:
It seems like it is good enough.

Fixes T1129

Test Plan:
Add a hex:forward .. at the top of %handler-wrap to wrap all requests, checked that https://staging.screenshotbot.io was loading from screenshotbot.io. (that's the first version of this diff.)

In the next version, I removed the hex:forward.. but made sure the api requests are still going through to the forwarded server.

Maniphest Tasks: T1129

Differential Revision: https://phabricator.tdrhq.com/D9361

GitOrigin-RevId: faf8363c17e802fbe68d1b6ab1a460383514bd15