Remove one level of indirection for SSO
Summary:
This undos D6085. This was only added to track how frequently
the OAuth was clicked, but it causes more problems than good for SSO,
especially on Enterprise.

In particular, look at the parent task for T1287: since these links
are exposed externally and when bots (for instance SlackBot trying to
get a screenshot of a page), open the page, rather than taking them to
a sign-in redirect, it takes them to a /n/ redirect that fails,
because no cookies are available.

Question: Does it make sense to undo D6084 too?

Maniphest Tasks: T1289

Differential Revision: https://phabricator.tdrhq.com/D9711

GitOrigin-RevId: b57850a84e1c0229feed970d862a2da9f75a7037