[sec] Reset session when logging in
Summary:
Fixes T904

Depends on D9779

Test Plan:
The biggest risk with this change is if an internal
session-value was set before the session was reset. The second biggest
risk was nibbles no longer being valid.

To test nibbles: Used the /joe link to sign-in using various flows,
both email and GitHub.

Maniphest Tasks: T904

Differential Revision: https://phabricator.tdrhq.com/D9780

GitOrigin-RevId: 8b1fb9ec3ccc425d350b3ce6387abb5641d11bf3