Remove :immediate-flush
Summary:
Could this be causing Gradle to misbehave?

It could be that the CLI tool is trying to flush, but Gradle isn't
pulling the output yet... The timing on the screenshot that Robin sent
might actually show the timing when the log was created, but not the
time that Gradle actually outputted the log.

Maniphest Tasks: T1514

Differential Revision: https://phabricator.tdrhq.com/D10235

GitOrigin-RevId: 4a9dbce6c14f26b6e5db09bda39a8a8fc3671f48