Use debian for OSS
Summary:
cimg/base uses an old version of Ubuntu with an old version
of SBCL that is crashing.

Test Plan:
Built this out on a PR directly on the OSS repo and then
copied this over.

This is sort of working, but one of the tests is crashing. We'll
attempt to fix that in the next step.

Differential Revision: https://phabricator.tdrhq.com/D10875

GitOrigin-RevId: c2847efd95cf8881117b6e73a217dbadda915527