Use the new flow even for "locally-rebased-p"
Summary:
This simplifies the commit-graph flow. In particular, in
conversation with ForeFlight, I found out that Bitrise provides an
easy way to run builds against the merge commit of a PR.

So they're not technically "rebasing" locally, this commit is actually
in the remote Git repository. It should not be a reason to stop using
the new flow.

Differential Revision: https://phabricator.tdrhq.com/D11747

GitOrigin-RevId: 3598002e1714d4bdef627e3398829c14b1516696