Fix deadlock when waiting for run from parent commit
Summary:
This is an old bug, and I should've caught it
earlier. Essentially we're waiting on the promotion-lock, but waiting
for run was using the channel-lock. This means, wait-for-run is never
going to finish promotion while we're waiting here.

I think we can simplify this further: when restarting, the promotion, we're actually waiting for a the promoted run on the channel to change.

Fixes T2138

Test Plan:
Repro-ed in fast-example:

Wrote two commits.

First ran on the second commit

Then went back and and ran on the first commit.

They're in deadlock now.

tried this after the fix, and it works

Maniphest Tasks: T2138

Differential Revision: https://phabricator.tdrhq.com/D11973

GitOrigin-RevId: 999f46cefbb5232bf0c3b02e0b05e614736d7fb2